commit 4 proyecto Maven

+++ proyectoClinica/src/main/java/com/backend/proyectoClinica/dto/salida/OdontologoSalidaDto.java
@@ -1,7 +1,7 @@
 package com.backend.proyectoClinica.dto.salida;
 
 public class OdontologoSalidaDto {
-    private int id;
+    private Long id;
     private int numeroMatricula;
     private String nombre;
     private String apellido;
@@ -9,18 +9,18 @@ public class OdontologoSalidaDto {
     public OdontologoSalidaDto() {
     }
 
-    public OdontologoSalidaDto(int id, int numeroMatricula, String nombre, String apellido) {
+    public OdontologoSalidaDto(Long id, int numeroMatricula, String nombre, String apellido) {
         this.id = id;
         this.numeroMatricula = numeroMatricula;
         this.nombre = nombre;
         this.apellido = apellido;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
