+++ proyectoClinica/src/main/java/com/backend/proyectoClinica/service/IPacienteService.java
@@ -10,5 +10,5 @@ public interface IPacienteService {
 
     List<PacienteSalidaDto> listarPacientes();
 
-    PacienteSalidaDto buscarPacientePorId(int id);
+    PacienteSalidaDto buscarPacientePorId(Long id);
 }
